# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Anno API
  description: |
    Anno - AI-Native Web Browser API

    Transform web content into AI-ready semantic streams with 82.3% token reduction.

    Anno provides intelligent web content extraction, semantic search, and RAG (Retrieval-Augmented Generation)
    capabilities optimized for AI agents and LLMs.
  version: 0.1.0
  contact:
    name: Anno Support
    url: https://github.com/evo-hydra/anno
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:5213
    description: Development server
  - url: https://api.anno.ai
    description: Production server

tags:
  - name: content
    description: Web content fetching and distillation
  - name: semantic
    description: Semantic search and RAG operations
  - name: memory
    description: Conversation memory management
  - name: health
    description: Health and monitoring endpoints

paths:
  /health:
    get:
      summary: Health check
      description: Returns detailed system health status including dependency checks
      tags:
        - health
      responses:
        '200':
          description: System is healthy or degraded
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'
        '503':
          description: System is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthStatus'

  /metrics:
    get:
      summary: Prometheus metrics
      description: Returns metrics in Prometheus exposition format
      tags:
        - health
      responses:
        '200':
          description: Metrics in Prometheus format
          content:
            text/plain:
              schema:
                type: string
                example: |
                  # HELP anno_fetch_total Total number of fetch requests
                  # TYPE anno_fetch_total counter
                  anno_fetch_total 42

  /v1/content/fetch:
    post:
      summary: Fetch and distill web content
      description: Fetches a URL and returns structured JSONL stream of semantic nodes
      tags:
        - content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - url
              properties:
                url:
                  type: string
                  format: uri
                  description: URL to fetch and process
                  example: https://example.com/article
                options:
                  type: object
                  properties:
                    useCache:
                      type: boolean
                      default: true
                      description: Use cached version if available
                    maxNodes:
                      type: integer
                      minimum: 1
                      maximum: 100
                      default: 40
                      description: Maximum number of semantic nodes to extract
                    render:
                      type: boolean
                      default: false
                      description: Use JavaScript rendering (Playwright)
      responses:
        '200':
          description: JSONL stream of semantic nodes
          content:
            application/x-ndjson:
              schema:
                type: string
                description: Newline-delimited JSON stream
              example: |
                {"type":"metadata","payload":{"url":"https://example.com"}}
                {"type":"node","payload":{"text":"Article content..."}}
                {"type":"done","payload":{"nodes":5}}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/content/batch-fetch:
    post:
      summary: Batch fetch multiple URLs in parallel
      description: Fetches multiple URLs concurrently and returns JSONL stream with results for each source
      tags:
        - content
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - urls
              properties:
                urls:
                  type: array
                  minItems: 1
                  maxItems: 10
                  items:
                    type: string
                    format: uri
                  description: Array of URLs to fetch (max 10)
                  example:
                    - https://example.com/article1
                    - https://example.com/article2
                options:
                  type: object
                  properties:
                    useCache:
                      type: boolean
                      default: true
                      description: Use cached version if available
                    maxNodes:
                      type: integer
                      minimum: 1
                      maximum: 100
                      default: 40
                      description: Maximum number of semantic nodes to extract per URL
                    render:
                      type: boolean
                      default: false
                      description: Use JavaScript rendering (Playwright)
                    parallel:
                      type: integer
                      minimum: 1
                      maximum: 5
                      default: 3
                      description: Number of URLs to fetch concurrently
      responses:
        '200':
          description: JSONL stream with batch events
          content:
            application/x-ndjson:
              schema:
                type: string
                description: |
                  Newline-delimited JSON stream with the following event types:
                  - batch_start: Signals start of batch processing
                  - source_start: Signals start of individual URL processing
                  - source_event: Wraps pipeline events (metadata, node, confidence, etc.) with source metadata
                  - source_end: Signals end of individual URL processing (success or error)
                  - batch_end: Signals end of batch processing
              example: |
                {"type":"batch_start","payload":{"totalUrls":2,"parallelism":3,"timestamp":1234567890}}
                {"type":"source_start","payload":{"url":"https://example.com","index":0,"timestamp":1234567890}}
                {"type":"source_event","payload":{"url":"https://example.com","index":0,"event":{"type":"metadata",...}}}
                {"type":"source_event","payload":{"url":"https://example.com","index":0,"event":{"type":"node",...}}}
                {"type":"source_end","payload":{"url":"https://example.com","index":0,"status":"success","timestamp":1234567891}}
                {"type":"batch_end","payload":{"totalUrls":2,"timestamp":1234567892}}
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/semantic/index:
    post:
      summary: Index documents
      description: Index documents into the semantic search vector store
      tags:
        - semantic
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - documents
              properties:
                documents:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required:
                      - id
                      - text
                    properties:
                      id:
                        type: string
                        description: Unique document identifier
                      text:
                        type: string
                        description: Document text for embedding
                      content:
                        type: string
                        description: Full document content
                      metadata:
                        type: object
                        additionalProperties: true
                        description: Additional metadata
      responses:
        '202':
          description: Documents indexed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: indexed
                  count:
                    type: integer
                    example: 5
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/semantic/search:
    post:
      summary: Semantic search
      description: Search indexed documents using semantic similarity
      tags:
        - semantic
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Search query
                  example: machine learning best practices
                k:
                  type: integer
                  minimum: 1
                  maximum: 20
                  default: 5
                  description: Number of results to return
                minScore:
                  type: number
                  minimum: 0
                  maximum: 1
                  description: Minimum similarity score threshold
                filter:
                  type: object
                  additionalProperties: true
                  description: Metadata filters
      responses:
        '200':
          description: Search results
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: '#/components/schemas/SearchResult'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/semantic/rag:
    post:
      summary: RAG (Retrieval-Augmented Generation)
      description: Perform semantic search and generate an AI-powered answer
      tags:
        - semantic
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - query
              properties:
                query:
                  type: string
                  description: Question to answer
                  example: What are the benefits of microservices?
                sessionId:
                  type: string
                  description: Conversation session ID for context
                k:
                  type: integer
                  minimum: 1
                  maximum: 10
                  default: 5
                  description: Number of documents to retrieve
                minScore:
                  type: number
                  minimum: 0
                  maximum: 1
                  description: Minimum relevance score
                summaryLevels:
                  type: array
                  items:
                    type: string
                    enum: [headline, paragraph, detailed]
                  description: Summary detail levels
                skipCache:
                  type: boolean
                  default: false
                  description: Skip query cache
      responses:
        '200':
          description: RAG response with answer and citations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RAGResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  /v1/memory/conversations/{sessionId}:
    get:
      summary: Get conversation history
      description: Retrieve conversation history for a session
      tags:
        - memory
      security:
        - ApiKeyAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          description: Session identifier
      responses:
        '200':
          description: Conversation history
          content:
            application/json:
              schema:
                type: object
                properties:
                  messages:
                    type: array
                    items:
                      $ref: '#/components/schemas/ConversationMessage'
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      summary: Clear conversation history
      description: Delete all messages for a session
      tags:
        - memory
      security:
        - ApiKeyAuth: []
      parameters:
        - name: sessionId
          in: path
          required: true
          schema:
            type: string
          description: Session identifier
      responses:
        '200':
          description: Session cleared
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: cleared
        '404':
          description: Session not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for authentication

  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
          description: Overall system status
        timestamp:
          type: integer
          format: int64
          description: Unix timestamp in milliseconds
        checks:
          type: object
          properties:
            cache:
              $ref: '#/components/schemas/ComponentHealth'
            ollama:
              $ref: '#/components/schemas/ComponentHealth'
            renderer:
              $ref: '#/components/schemas/ComponentHealth'
        overall:
          type: object
          properties:
            healthy:
              type: integer
            degraded:
              type: integer
            unhealthy:
              type: integer

    ComponentHealth:
      type: object
      properties:
        status:
          type: string
          enum: [healthy, degraded, unhealthy]
        message:
          type: string
        latencyMs:
          type: number
        details:
          type: object
          additionalProperties: true

    SearchResult:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        score:
          type: number
          minimum: 0
          maximum: 1
        metadata:
          type: object
          additionalProperties: true

    RAGResponse:
      type: object
      properties:
        answer:
          type: string
          description: Generated answer
        citations:
          type: array
          items:
            type: object
            properties:
              documentId:
                type: string
              text:
                type: string
              score:
                type: number
        summaries:
          type: array
          items:
            type: object
            properties:
              level:
                type: string
                enum: [headline, paragraph, detailed]
              text:
                type: string
              confidence:
                type: number
        _cached:
          type: boolean
          description: Whether result was served from cache

    ConversationMessage:
      type: object
      properties:
        role:
          type: string
          enum: [user, assistant, system]
        content:
          type: string
        timestamp:
          type: integer
          format: int64

    Error:
      type: object
      properties:
        error:
          type: string
          description: Error code
        message:
          type: string
          description: Human-readable error message
        statusCode:
          type: integer
          description: HTTP status code
        details:
          type: object
          additionalProperties: true
          description: Additional error details
        timestamp:
          type: integer
          format: int64
        path:
          type: string
          description: Request path that caused the error